refine structure

+++ scripts/install-symlink.sh
@@ -5,12 +5,18 @@ set -u
 PARENT_DIR=$(cd $(dirname $0); cd ..; pwd)
 FILES=(
     ".dir_colors"
-    ".gitignore.global"
     ".inputrc"
     ".nanorc"
+    ".git.d"
     ".zsh.d"
 )
 
 for f in ${FILES[@]}; do
-    ln -s $PARENT_DIR/$f $HOME/$f
+    # if symlink to directory already exists
+    # `ln` creates symlink on linked directory
+    if [ -e $HOME/$f ]; then
+        echo "$HOME/$f already exists"
+    else
+        ln -s $PARENT_DIR/$f $HOME/$f
+    fi
 done
